Stop kludging file extension into classifier
Previous approach blindly appended ".asc" to the classifier, breaking resolution for special artifacts like test JARs. Maven resolves these artifacts to a packaging type and file extension of "jar", but wasn't doing this here because the classifiers didn't match (since it had ".asc" at the end).

New approach provides the original classifier to Maven for it to apply standard look-up rules, then uses a new AscArtifactHandler that decorates the ArtifactHandler Maven returns for the artifact so that we fetch an artifact ending in ".asc".

Closes #22.