[#25] Stop using .properties files for keys
the format being used is not quite that of a properties file. the fact that CheckStyle is unhappy with a sample keys map "properties" file that is accepted by PGP confirms this. I've renamed the vars, changed the docs, and updated the sample files. the new file extension is ".list" for keys map files.