Don't Verify Reactor Deps by Default (#29)
this change adjusts PGP Verify to no longer complain about dependencies that are being built as part of the current multi-module build, in the event that the current build is not being signed. this allows `mvn clean install` to work properly without requiring that every build be signed, as long as the dependencies of all of the projects in the current build (minus the projects currently being built) are signed. the option can be toggled on or off with the new `verifyReactorDependencies` setting on the plug-in, in the event that projects want to require signing for both dependencies AND projects in the current reactor build.