[#34] Switch to the Apache HTTP Client
Java's built-in HTTP client does not provide automatic retry behavior, nor does it seem as robust as Apache's HTTP client library.

This re-factor changes the plug-in to use the Apache client, as follows:
- Implements a retry strategy to allow PGP key requests to recover automatically on transient 408 errors, 5XX errors, connection timeouts, and read timeouts.
- Sets the default timeouts to:
  - 5 seconds to connect to the target server
  - 10 seconds to read the key from the server
- Implements a back-off strategy to allow unavailable servers to recover by adding +750 ms each time we retry a connection to an unreachable server.

Timeouts and back-offs are not yet user-configurable, but can be made configurable at a later time should the need arise.

Closes #34.