Add 'strictNoSignature' which expects artifacts without signature in the keys map.
In order to detect and support the use cases "some version of artifact
comes without signature" and "strict preselection of artifacts having
no signature", the property 'strictNoSignature' will refer to the keys
map to verify whether or not having no signature is acceptable for
each unsigned artifact.

This mode is less strict than 'failNoSignature' as it will be able to
accomodate artifacts without signature. At the same time it provides
explicit control over which artifacts are acceptable without signature.