Different mechanism for resolving dependencies and plugins.
Now using a different mechanism to resolve dependencies and plugins.
This mechanism seems to be the intended way in Maven. It does not,
however, resolve build plug-in dependencies. This is not implemented at
this stage and will need to be added later.

The dependency and plugin artifacts already have resolved the leading
artifact version. We do perform additional filtering according to
configuration.