Fix 'noKey' specifier for PGP keysmap.
The control flow logic for the 'noKey' specifier is broken as it matches
on the wrong exception to identify the case. This is fixed and while
fixing starts using Java's catch-syntax capabilities for implicitly
performing the instance-of test.

Added IT tests 'noKeyOK' and 'noKeyFail' for verifying correct behavior.
Both tests use a PGP keypair that was never published. The 'helloworld'
artifact is installed in the local repo used for IT tests only.

Slightly modified capitalization in the exception's error message
because of my OCD.