Introduce checksum to skip repeated runs.
Introduce checksum for the artifact list such that there is a way to
determine whether or not anything has changed in the set of artifacts
that need to be validated.
A SHA-256 digest is calculated over all resolved artifacts, i.e. their
full identities, that participate in the pgpverify artifact validation
and compared against the digest that was saved in a previous run. The
set of artifacts are expected to have a definite order determined by the
order of resolution.
If the checksum of a previous run and of the current run are the same,
it is safe to assume that the set of artifacts for the project has not
changed, therefore we can assume that if a previous validation succeeded
that it will succeed for this run too.

As the checksum is calculated from the set of all artifacts that are
resolved, changes to the pgpverify configuration may result in a
different set of artifacts, thus a new full validation run.
The configuration parameter `disableChecksum` allows for disabling the
use of the checksum to exit early and instead do a full validation run
every time.