Added regular expression matching for artifact version
When the version field begins with ` ~ ` (tilde), everything following the tilde is a case-insensitive regular
expression matched against artifact version.

When the version field begins with ` !~ ` (not-tilde), everything following the not-tilde is a case-insensitive regular
expression with a negated match against artifact version.

This may be used, for example, to have separate PGP signing keys for continuous integration snapshot builds,
while releases are signed by more protected PGP keys unavailable to the continuous integration platform.